Invert dependency from Logic to Storage [LO-DIP]
